refactor: status_cache 構造分割と merge_readiness キャッシュ設計修正
- status_cache/mod.rs を status_cache.rs に変更（Rust 2018+ スタイル）
- status_cache/domain.rs を domain/cache.rs と domain/daemon.rs に分割
- merge_readiness::domain::cache と application::cache を削除
  （キャッシュはドメイン概念ではなくインフラ関心事）
- DaemonCacheAdapter と run_cached_prompt を main.rs から除去
- merge_readiness::interface::cli::prompt::cached を追加
  （キャッシュモードの CLI 実装を正しい層に配置）
- status_cache::application::cache::query_via_daemon を追加
  （DaemonClient をアプリケーション層で隠蔽）
