+++ src/contexts/merge_readiness/domain.rs
@@ -1,5 +1,4 @@
 pub mod branch_sync;
-pub mod cache;
 pub mod ci_checks;
 pub mod error;
 pub mod merge_ready;
