--- src/contexts/merge_readiness/domain/cache.rs
@@ -1,9 +1,0 @@
-pub enum CacheState {
-    Fresh(String),
-    Stale(String),
-    Miss,
-}
-
-pub trait CachePort {
-    fn check(&self, repo_id: &str) -> CacheState;
-}
